refactor: split code to files
